Updated as per pr review (NOTE: this is a rough implementation to see if the build system would work)
