+++ obc/CMakeLists.txt
@@ -107,6 +107,7 @@ target_compile_definitions(${OUT_FILE_NAME} PRIVATE
     CSDC_DEMO_ENABLED=${CSDC_DEMO_ENABLED}
     ENABLE_TASK_STATS_COLLECTOR=${ENABLE_TASK_STATS_COLLECTOR}
     ENABLE_BL_BYPASS=${ENABLE_BL_BYPASS}
+    IS_BL=0
     ${PERIPHERAL_CONFIG}
 )
 
@@ -202,6 +203,7 @@ target_compile_definitions(OBC-bl.out PRIVATE
     OBC_UART_BAUD_RATE=${OBC_UART_BAUD_RATE}
     CSDC_DEMO_ENABLED=${CSDC_DEMO_ENABLED}
     ENABLE_TASK_STATS_COLLECTOR=${ENABLE_TASK_STATS_COLLECTOR}
+    IS_BL=0
     ${PERIPHERAL_CONFIG}
 )
 
@@ -233,6 +235,7 @@ target_compile_definitions(debug-tool.out PRIVATE
     OBC_UART_BAUD_RATE=${OBC_UART_BAUD_RATE}
     CSDC_DEMO_ENABLED=${CSDC_DEMO_ENABLED}
     ENABLE_TASK_STATS_COLLECTOR=${ENABLE_TASK_STATS_COLLECTOR}
+    IS_BL=0
     ${PERIPHERAL_CONFIG}
 )
