+++ obc/app/modules/command_mgr/command_manager.c
@@ -1,9 +1,10 @@
+#include "alarm_handler.h"
+#include "command.h"
 #include "command_manager.h"
 #include "obc_gs_command_data.h"
 #include "obc_errors.h"
 #include "obc_logging.h"
 #include "obc_assert.h"
-#include "command.h"
 
 #include <FreeRTOS.h>
 #include <sys_common.h>
@@ -16,8 +17,6 @@
 static QueueHandle_t commandQueueHandle;
 static StaticQueue_t commandQueue;
 static uint8_t commandQueueStack[COMMAND_QUEUE_LENGTH * COMMAND_QUEUE_ITEM_SIZE];
-
-STATIC_ASSERT(CMDS_CONFIG_SIZE <= UINT8_MAX, "Max command ID must be less than 256");
 
 void obcTaskInitCommandMgr(void) {
   ASSERT((commandQueueStack != NULL) && (&commandQueue != NULL));
@@ -43,22 +42,42 @@ obc_error_code_t sendToCommandQueue(cmd_msg_t *cmd) {
   return OBC_ERR_CODE_QUEUE_FULL;
 }
 
-void obcTaskFunctionCommandMgr(void *pvParameters) {
+obc_error_code_t processTimeTaggedCommand(cmd_msg_t *cmd, cmd_info_t *currCmdInfo) {
+  // If the timetag is in the past, throw away the command
   obc_error_code_t errCode;
+  if (!cmd->isTimeTagged) {
+    return OBC_ERR_CODE_INVALID_ARG;
+  }
+  if (cmd->timestamp < getCurrentUnixTime()) {
+    return OBC_ERR_CODE_SUCCESS;
+  }
 
-  // Used to track whether a safety-critical command is currently being executed
-  // This is inefficient space-wise, but simplifies the code. We can optimize later if needed.
-  static bool cmdProgressTracker[sizeof(cmdsConfig) / sizeof(cmd_info_t)] = {false};
+  alarm_handler_event_t alarm = {.id = ALARM_HANDLER_NEW_ALARM,
+                                 .alarmInfo = {
+                                     .unixTime = cmd->timestamp,
+                                     .callbackDef =
+                                         {
+                                             .cmdCallback = currCmdInfo->callback,
+                                         },
+                                     .type = ALARM_TYPE_TIME_TAGGED_CMD,
+                                     .cmdMsg = *cmd,
+                                 }};
+
+  RETURN_IF_ERROR_CODE(sendToAlarmHandlerQueue(&alarm));
+  return OBC_ERR_CODE_SUCCESS;
+}
+
+void obcTaskFunctionCommandMgr(void *pvParameters) {
+  obc_error_code_t errCode = 0;
 
   while (1) {
     cmd_msg_t cmd;
     if (xQueueReceive(commandQueueHandle, &cmd, portMAX_DELAY) == pdPASS) {
       cmd_info_t currCmdInfo;
 
-      errCode = verifyCommand(&cmd, &currCmdInfo, cmdProgressTracker);
+      LOG_ERROR_CODE(verifyCommand(&cmd, &currCmdInfo));
 
       if (errCode != OBC_ERR_CODE_SUCCESS) {
-        LOG_ERROR_CODE(errCode);
         continue;
       }
