+++ obc/shared/commands/command.c
@@ -1,10 +1,28 @@
 #include "command.h"
 #include "obc_logging.h"
-#include "alarm_handler.h"
+
+extern const cmd_info_t cmdConfig[];
+
+#ifdef IS_BL
+static const cmd_info_t cmdsConfig[] = {[CMD_PING] = {pingCmdCallback, CMD_POLICY_PROD, CMD_TYPE_NORMAL}};
+#else
+static const cmd_info_t cmdsConfig[] = {
+    [CMD_END_OF_FRAME] = {NULL, CMD_POLICY_PROD, CMD_TYPE_NORMAL},
+    [CMD_EXEC_OBC_RESET] = {execObcResetCmdCallback, CMD_POLICY_PROD, CMD_TYPE_CRITICAL},
+    [CMD_RTC_SYNC] = {rtcSyncCmdCallback, CMD_POLICY_PROD, CMD_TYPE_NORMAL},
+    [CMD_DOWNLINK_LOGS_NEXT_PASS] = {downlinkLogsNextPassCmdCallback, CMD_POLICY_PROD, CMD_TYPE_CRITICAL},
+    [CMD_MICRO_SD_FORMAT] = {microSDFormatCmdCallback, CMD_POLICY_PROD, CMD_TYPE_CRITICAL},
+    [CMD_PING] = {pingCmdCallback, CMD_POLICY_PROD, CMD_TYPE_NORMAL},
+    [CMD_DOWNLINK_TELEM] = {downlinkTelemCmdCallback, CMD_POLICY_PROD, CMD_TYPE_NORMAL}};
+#endif
 
 #define CMDS_CONFIG_SIZE (sizeof(cmdsConfig) / sizeof(cmd_info_t))
 
-obc_error_code_t verifyCommand(cmd_msg_t *cmd, cmd_info_t *currCmdInfo, bool cmdProgressTracker[]) {
+// Used to track whether a safety-critical command is currently being executed
+// This is inefficient space-wise, but simplifies the code. We can optimize later if needed.
+static bool cmdProgressTracker[sizeof(cmdsConfig) / sizeof(cmd_info_t)] = {false};
+
+obc_error_code_t verifyCommand(cmd_msg_t *cmd, cmd_info_t *currCmdInfo) {
   if (cmd->id >= CMDS_CONFIG_SIZE) {
     return OBC_ERR_CODE_UNSUPPORTED_CMD;
   }
@@ -45,30 +63,5 @@ obc_error_code_t processNonTimeTaggedCommand(cmd_msg_t *cmd, cmd_info_t *currCmd
   }
   // TODO: Handle safety-critical command failures
   RETURN_IF_ERROR_CODE(currCmdInfo->callback(cmd));
-  return OBC_ERR_CODE_SUCCESS;
-}
-
-obc_error_code_t processTimeTaggedCommand(cmd_msg_t *cmd, cmd_info_t *currCmdInfo) {
-  // If the timetag is in the past, throw away the command
-  obc_error_code_t errCode;
-  if (!cmd->isTimeTagged) {
-    return OBC_ERR_CODE_INVALID_ARG;
-  }
-  if (cmd->timestamp < getCurrentUnixTime()) {
-    return OBC_ERR_CODE_SUCCESS;
-  }
-
-  alarm_handler_event_t alarm = {.id = ALARM_HANDLER_NEW_ALARM,
-                                 .alarmInfo = {
-                                     .unixTime = cmd->timestamp,
-                                     .callbackDef =
-                                         {
-                                             .cmdCallback = currCmdInfo->callback,
-                                         },
-                                     .type = ALARM_TYPE_TIME_TAGGED_CMD,
-                                     .cmdMsg = *cmd,
-                                 }};
-
-  RETURN_IF_ERROR_CODE(sendToAlarmHandlerQueue(&alarm));
   return OBC_ERR_CODE_SUCCESS;
 }
